♻️ restructured modules for new gateset structure
